pyproject.toml
[project]
name = "myblog"
version = "0.1.0"
description = "Blog platform REST API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-filter>=23.5",
    "drf-spectacular>=0.27",
    "psycopg2-binary>=2.9",
    "requests>=2.31",
    "django-cors-headers>=4.3",
    "dj-database-url>=0.5.0",
    "django-ckeditor>=6.7.3",
    "uvicorn>=0.44.0",
    "whitenoise>=6.12.0",
    "python-dotenv>=1.2.2",
]
